rag/requirements_rag.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
openai
pandas
pathlib
pymupdf
python-dotenv